addProudctAddWithoutloadfile

+++ MyTestAngular/Controllers/TypeController.cs
@@ -21,20 +21,20 @@ namespace MyTestAngular.Controllers
         }
 
         [HttpGet]
-        public IEnumerable<ApplicationType> Get()
+        public async Task<IEnumerable<ApplicationType>> Get()
         {
-            return _db.ApplicationTypes;
+            return await _db.ApplicationTypes.ToListAsync();
         }
 
         [HttpPost]
-        public void Add(ApplicationType type)
+        public async Task<ActionResult<ApplicationType>> Add(ApplicationType type)
         {
-            if (type != null)
-            {
-                _db.ApplicationTypes.Add(type);
-                _db.SaveChanges();
-            }
+            if (type == null)
+                return BadRequest();
 
+            _db.ApplicationTypes.Add(type);
+            await _db.SaveChangesAsync();
+            return Ok(type);
         }
 
         [HttpPut]
@@ -49,26 +49,25 @@ namespace MyTestAngular.Controllers
                 return NotFound();
 
             objCat.Name = type.Name;
-
             await _db.SaveChangesAsync();
 
             return Ok(type);
 
         }
 
-
-
         [HttpDelete("{id}")]
-        public ActionResult Delete([FromRoute]int id)
+        public async Task<ActionResult<ApplicationType>> Delete([FromRoute]int id)
         {
+            if(id==null)
+                return BadRequest();
 ;
             var obj = _db.ApplicationTypes.Find(id);
             if (obj == null)
-            {
                 return NotFound();
-            }
+    
+
             _db.ApplicationTypes.Remove(obj);//Удаление в бд
-            _db.SaveChanges();// Сохранение в бд - только после этого категория попадет в бд
+            await _db.SaveChangesAsync();// Сохранение в бд - только после этого категория попадет в бд
             return Ok();
 
         }
